Rearrange notificationprofile tests
